Move logic to FastlyHeaders controller concern
- Only add cache-control headers
